refactor(#97): presign url 제공하는 로직 개선
* 저수준 컴포넌트 -> 고수준 컴포넌트 의존을 변경
  * 저수준 컴포넌트 -> 저수준 컴포넌트
* 너무 과하게 분리된 클래스를 통합하면서 응집성 향상
