+++ src/main/java/deepdivers/community/domain/uploader/ui/UploaderOpenController.java
@@ -1,10 +1,11 @@
 package deepdivers.community.domain.uploader.ui;
 
 import deepdivers.community.domain.common.API;
-import deepdivers.community.domain.uploader.application.UploaderService;
 import deepdivers.community.domain.uploader.application.dto.request.GetPresignRequest;
 import deepdivers.community.domain.uploader.application.dto.response.GetPresignResponse;
+import deepdivers.community.domain.uploader.application.dto.response.statustype.UploaderStatusType;
 import deepdivers.community.domain.uploader.ui.docs.UploaderOpenControllerDocs;
+import deepdivers.community.infra.aws.s3.S3PresignManager;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -18,16 +19,21 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/open/uploader")
 public class UploaderOpenController implements UploaderOpenControllerDocs {
 
-    private final UploaderService uploaderService;
+    private final S3PresignManager s3PresignManager;
 
     @PostMapping("/presigned-url-generation")
     public ResponseEntity<API<GetPresignResponse>> generatePresignUrl(
         @Valid @RequestBody
         final GetPresignRequest request
     ) {
-        return ResponseEntity.ok(
-            uploaderService.generatePresignUrl(request)
-        );
+        final String key = s3PresignManager.generateKey(request.contentType(), request.keyType());
+        final String presignedUrl = s3PresignManager.generatePreSignedUrl(key, request.contentType());
+        final String accessUrl = s3PresignManager.generateAccessUrl(key);
+
+        return ResponseEntity.ok(API.of(
+            UploaderStatusType.GENERATE_PRESIGN_SUCCESS,
+            GetPresignResponse.of(key, presignedUrl, accessUrl)
+        ));
     }
 
 }
