+++ src/main/java/deepdivers/community/infra/aws/s3/S3PresignManager.java
@@ -1,11 +1,10 @@
 package deepdivers.community.infra.aws.s3;
 
-import static deepdivers.community.infra.aws.s3.generator.S3RequestGenerator.generatePresignRequest;
-import static deepdivers.community.infra.aws.s3.generator.S3RequestGenerator.generatePutObjectRequest;
 import static deepdivers.community.infra.aws.s3.validator.S3FileValidator.validateContentType;
 
 import deepdivers.community.infra.aws.s3.generator.S3KeyGenerator;
 import deepdivers.community.infra.aws.s3.properties.S3Properties;
+import java.time.Duration;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
@@ -16,6 +15,9 @@ import software.amazon.awssdk.services.s3.presigner.model.PutObjectPresignReques
 @Component
 @RequiredArgsConstructor
 public class S3PresignManager {
+
+    private static final Duration PRE_SIGNED_URL_EXPIRATION = Duration.ofMinutes(5);
+    private static final String TAG_INITIAL = "Status=Deleted";
 
     private final S3Properties s3Properties;
     private final S3Presigner s3Presigner;
@@ -41,6 +43,26 @@ public class S3PresignManager {
 
     public String generateAccessUrl(final String key) {
         return String.format("%s/%s", s3Properties.getBaseUrl(), key);
+    }
+
+    public static PutObjectPresignRequest generatePresignRequest(final PutObjectRequest objectRequest) {
+        return PutObjectPresignRequest.builder()
+            .signatureDuration(PRE_SIGNED_URL_EXPIRATION)
+            .putObjectRequest(objectRequest)
+            .build();
+    }
+
+    public static PutObjectRequest generatePutObjectRequest(
+        final String key,
+        final String contentType,
+        final String bucket
+    ) {
+        return PutObjectRequest.builder()
+            .bucket(bucket)
+            .key(key)
+            .contentType(contentType)
+            .tagging(TAG_INITIAL)
+            .build();
     }
 
 }
