+++ src/main/java/deepdivers/community/infra/aws/s3/exception/S3Exception.java
@@ -13,4 +13,5 @@ public enum S3Exception implements ExceptionType {
 
     private final int code;
     private final String message;
+
 }
